backend/requirements.txt
fastapi
uvicorn
pydantic-settings
requests
sqlalchemy
psycopg2-binary
python-multipart
pycryptodome
pgvector
pyodbc
openpyxl
